Move database initialization SQL scripts
There doesn't appear to be any reason that these SQL scripts exist in
the includes/ directory, so they've been moved to a location with a
more pertinent name. This directory is also used by the database docker
container to run database initialization scripts, so it may make more
sense to separate them from the rest of whatever is in include (which
doesn't get run, as that container's init script doesn't run PHP files).
This also restores numbering to the file names of these init SQL
scripts, which got removed in commit 5f1a4943. These are required as
the database container's initialization script runs scripts in a sorted
order, and we need to create all the tables before inserting an admin
account.