Remove comments about Docker volumes
The MySQL Docker image is already able to handle only running the
initialization scripts when the database files don't exist (i.e. when
`/var/lib/mysql`, mapped to `./data`, is empty), so after the first run
and `./data` is populated we don't need to manually modify the
`docker-compose.yml` file.